Remove dependence on two internal forks
_curator.go_

The original project was abandoned and I've migrated us to the fork
https://github.com/curator-go/curator. That has consumed the code changes david
had originally made to fix some bugs.

_go-metrics_

Upstream has consumed all the changes we made so I'm just moving us back to
upstream at latest.

In order to do the upgrade I had to update a few libraries, but they are all
libraries used to test. The tests continue to pass, so this should be ok.

I also upgraded the travis.yml file to use go1.8 as I had neglected to
previously.